File: SRC/TRF3dV100/LUgpuCHandle_interface_impl.cu Refactor LUgpuCHandle_interface_impl.cu
Files Changed:
- LUgpuCHandle_interface_impl.cu

Detailed Changes:
- Removed commented out code that was not being used.
- Removed unnecessary return statement.
- Added functions for handling single precision LUgpu handles.
- Added function for destroying LUgpu handles.
- Added function for copying LUgpu data from GPU to host.
- Added function for performing LU factorization in single precision.

File: SRC/TRF3dV100/lupanels_GPU_impl.hpp
Refactor GPU implementation in lupanels_GPU_impl.hpp and xupanels_GPU_impl.hpp

Files Changed:
- lupanels_GPU_impl.hpp: Refactored panelSolveGPU function to use myCublasTrsm instead of cublasDtrsm. Refactored diagFactorCuSolver function to use myCusolverGetrf instead of cusolverDnDgetrf.
- xupanels_GPU_impl.hpp: Refactored panelSolveGPU function to use myCublasTrsm instead of cublasDtrsm.

Detailed Changes:
- lupanels_GPU_impl.hpp:
  - Refactored panelSolveGPU function to use myCublasTrsm instead of cublasDtrsm. Changed parameters: CUBLAS_SIDE_RIGHT, CUBLAS_FILL_MODE_UPPER, CUBLAS_OP_N, CUBLAS_DIAG_NON_UNIT.
  - Refactored diagFactorCuSolver function to use myCusolverGetrf instead of cusolverDnDgetrf. Changed parameters: kSupSize, kSupSize.

- xupanels_GPU_impl.hpp:
  - Refactored panelSolveGPU function to use myCublasTrsm instead of cublasDtrsm. Changed parameters: CUBLAS_SIDE_LEFT, CUBLAS_FILL_MODE_LOWER, CUBLAS_OP_N, CUBLAS_DIAG_UNIT.

File: SRC/TRF3dV100/lupanels_impl.hpp
Summary: Refactor xLUstruct_t class and add new functions for initialization and computation.

Files Changed:
1. lupanels_impl.hpp

Detailed Changes:
- Added a new function `initDiagFactBufsArr` to initialize `diagFactBufs` array.
- Added a new function `getBigV` to allocate memory for `bigV` array.
- Commented out the line `bigV = dgetBigV(ldt, nThreads);` and replaced it with `bigV = getBigV<Ftype>(ldt, nThreads);`.
- Modified the constructor of `xLUstruct_t` class to include new parameters and initialize variables.
- Resized `diagFactBufs` vector to `numDiagBufs`.
- Commented out the line `dFBufs = dinitDiagFactBufsArr(mxLeafNode, ldt, grid);` and replaced it with `dFBufs = initDiagFactBufsArr(numDiagBufs, ldt);`.
- Modified the function `packedU2skyline` to include a new parameter of type `LUStruct_type<Ftype>`.
- Modified the function `blockUpdate` to use the `superlu_gemm` function template with the appropriate data type.

File: SRC/TRF3dV100/schurCompUpdate_impl.cuh
Refactor GPU matrix multiplication in schurCompUpdate_impl.cuh

Files Changed:
- schurCompUpdate_impl.cuh: Refactored GPU matrix multiplication using `myCublasGemm` wrapper.
- cublas_cusolver_wrappers.hpp: Added `myCublasGemm` function.

Detailed Changes:
- Replaced `cublasDgemm` calls with `myCublasGemm` calls in `xLUstruct_t<Ftype>::dSchurComplementUpdateGPU` and `xLUstruct_t<Ftype>::dSchurCompUpdatePartGPU` functions.
- Added `#include "cublas_cusolver_wrappers.hpp"` to `schurCompUpdate_impl.cuh` file.

File: SRC/TRF3dV100/xlupanels.hpp
Refactor xLUstruct_t in xlupanels.hpp

Files Changed:
- xlupanels.hpp

Detailed Changes:
- Replaced dtrf3Dpartition_t with trf3dpartitionType<Ftype> in xLUstruct_t constructor
- Replaced dtrf3Dpartition_t with trf3dpartitionType<Ftype> in xLUstruct_t definition
- Added initDiagFactBufsArr() function to initialize diagFactBufsArr
- Added marshallBatchedLUData() function to marshal batch LU data into A_gpu
- Added marshallBatchedBufferCopyData() function to marshal batch buffer copy data
- Added packedU2skyline() function to convert packed U to skyline format
- Added ancestorReduction3d() function for ancestor reduction in 3D

File: SRC/include/superlu_summit.h
Refactor sCreateLUgpuHandle function in superlu_summit.h

Files Changed:
- SRC/include/superlu_summit.h

Detailed Changes:
- Commented out the previous declaration of sCreateLUgpuHandle function.
- Added a new declaration of sCreateLUgpuHandle function with a modified parameter type for 'thresh_' from double to float.

+++ SRC/include/superlu_summit.h
@@ -63,11 +63,15 @@ typedef struct cLUstruct_t cLUstruct_t;
     struct sLUstruct_v1;
     typedef struct sLUstruct_v1 *sLUgpu_Handle;
 
-    extern sLUgpu_Handle sCreateLUgpuHandle(int_t nsupers, int_t ldt_, 
-    strf3Dpartition_t *trf3Dpartition,
-                                           sLUstruct_t *LUstruct, gridinfo3d_t *grid3d,
-                                           SCT_t *SCT_, superlu_dist_options_t *options_, SuperLUStat_t *stat,
-                                           double thresh_, int *info_);
+    // extern sLUgpu_Handle sCreateLUgpuHandle(int_t nsupers, int_t ldt_, 
+    // strf3Dpartition_t *trf3Dpartition,
+    //                                        sLUstruct_t *LUstruct, gridinfo3d_t *grid3d,
+    //                                        SCT_t *SCT_, superlu_dist_options_t *options_, SuperLUStat_t *stat,
+    //                                        double thresh_, int *info_);
+    extern sLUgpu_Handle sCreateLUgpuHandle(int_t nsupers, int_t ldt_, strf3Dpartition_t *trf3Dpartition,
+                                     sLUstruct_t *LUstruct, gridinfo3d_t *grid3d,
+                                     SCT_t *SCT_, superlu_dist_options_t *options_, SuperLUStat_t *stat,
+                                     float thresh_, int *info_);
 
     extern void sdestroyLUgpuHandle(sLUgpu_Handle LuH);
